requirements.txt
~streamlit
langchain-groq
langchain-community
langchain-huggingface
langchain-text-splitters
sentence-transformers
chromadb
pypdf
python-dotenv
tiktoken